df(1): fix for explicit non-mountpoint targets
It is not enough to simply match the path against what is given,
as that will fail if what is given is a path to a file and not
a mountpoint. We need to get the device for each mountpoint
and then try to match that against the device of the file (which
is already stat'd).