chmod: fix recursive version for symlinks
On FreeBSD it is possible to change the mode of a symlink, so
fchmodat is used, with AT_SYMLINK_NOFOLLOW. On Linux, we cannot
change the mode of a symlink, and trying to use the flag with
fchmodat results in ENOTSUP.

The desired behavior is to just not do anything to symlinks,
which is how other Linux chmod implementations behave. We already
know when the file is a symlink (this information is given from
fts), so just use regular chmod for non-symlinks or when it is
requested that we dereference links. There was no other purpose
to using fchmodat instead of chmod other than being able to pass
the AT_SYMLINK_NOFOLLOW flag (the behavior is otherwise identical
with AT_FDCWD).