whereis: skip if we can't query manpage paths
Also use man -w instead of manpath, as with mandoc manpath does
not exist.

Also do not fail if we can't query the paths, it may be just so
that e.g. base-man in chimera is not installed, and we still
want to print other results (i.e. pretend there are no manpaths).