seq(1): drop a workaround in seq(1) causing bad number of lines
This would previously result in behavior like:

$ seq 1 1050000|wc -l
 1050001

The reason being something like this:

$ seq 1050000 1050000
1.05e+06
1.05e+06

The source of the issue is the following commit:

https://github.com/freebsd/freebsd-src/commit/3049d4ccc031e5d4d0ff6b9a2445131f507778f2

The problem is that 'cur' and 'last_shown_value' will never not
be not equal, as 'cur' is always incremented past the 'last'
boundary in order to break the loop. As far as I can tell, there
is no reliable way to check if there was a bad increment due to
rounding error, so we should probably disregard those cases.

The checking of whether we were dealing with integers is not
reliable here either. The default %g format will truncate for
any value over 999999. The double type can represent a lot larger
range of integer values than that; treating them as truncating
floats is wrong.

Revert for now, until a better solution can be found.