seq(1): print integers up to 2^mantissa as integers
The previous behavior would unconditionally use the %g format
string by default, which resulted in all integers above 999999
being truncated and printed in exponent form.

As this behavior differs from GNU seq(1) and is not very useful
in most cases, perform additional checks for whether all three
given numbers are really integers (i.e. they do not truncate if
floored and their absolute value can be represented as an integer
within mantissa precision) and if that is the case, print the
whole sequence as if it was integers.

If any part requires floating point, the old behavior is still
followed. The existing behavior of zero-or-space padding is also
followed, merely adapting it for integer printing.

Additionally, for when we are in charge of the format string and
we know that no formatting truncation will happen, skip past the
workaround for last value of truncated prints in order to avoid
memory allocations, for more robust behavior.