add custom realpath/readlink implementation
This implementation is somewhere between the BSD realpath and
GNU realpath. Like the BSD realpath, it always requires the
path to exist by default (mirroring the realpath(3) C API)
while also providing a variant that does not require anything
to exist; the "only last component may not exist" default mode
is not provided. It also does not implement logical .. expansion
and relative path printing; these may be added later but for now
they would complicate the implementation and possibly be subject
to races.

Like the BSD realpath tool, this is a relatively thin wrapper
around C++ std::filesystem and its operations.

While at it, it implements GNU-style readlink functionality,
replacing the one provided by stat if needed.

Manpages are still pending. This is for now not the default.