find: fix compatibility with GNU getopt(3)
By default, GNU getopt(3) will continue parsing options after the
first nonoption. This behavior is incompatible with this find(1)
implementation, causing parts of the expression to be parsed as
(invalid) options.

The behavior can be disabled by prefixing the option string with
a `+` character. Other applets, such as flock or mcookie do it
already so it would be cool to do this here too to make find(1)
compatible with glibc.