sh(1): ignore extra arguments for command -v foo [args]
The FreeBSD sh errors by default, however dash just ignores the
extra arguments (which is convenient sometimes). Bash seems weird
here as it considers each word a separate command and succeeds if
at least one of them resolves, however this is at odds with the
official documentation, which specifies the extra arguments are
really just arguments to command and only the first command arg
should be evaluated.

Follow the dash behavior for best compatibility.